将查询归纳到一个action，不知道会不会有问题，今天才领略到Hibernate在初始化时有多慢，查询都能费上3s...这还怎么玩。。。看来在登录时得依次预加载下

+++ src/cn/hibernate/beans/StudentDAO.java
@@ -59,7 +59,7 @@ public class StudentDAO extends BaseHibernateDAO {
 		}
 	}
 
-	public Student findById(java.lang.Integer id) {
+	public Student findById(java.lang.Long id) {
 		log.debug("getting Student instance with id: " + id);
 		try {
 			Student instance = (Student) getSession().get(
